added toggling between city search and zip code search, began work on city search
