+++ zip-search/src/App.css
@@ -1,6 +1,3 @@
-.App-header {
+.body-bg-color{
   background-color: #222;
-  padding: 20px;
-  color: white;
-  text-align: center;
 }
