+++ zip-search/src/zip-search.css
@@ -1,0 +1,6 @@
+.App-header {
+  background-color: #222;
+  padding: 20px;
+  color: white;
+  text-align: center;
+}
